Moving query assert context manager to its own module
